+++ pkg/monitoring/alloy/configmap_test.go
@@ -10,12 +10,10 @@ import (
 	"github.com/blang/semver/v4"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/utils/ptr"
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 
-	"github.com/giantswarm/observability-operator/pkg/common"
 	commonmonitoring "github.com/giantswarm/observability-operator/pkg/common/monitoring"
-	"github.com/giantswarm/observability-operator/pkg/monitoring"
+	"github.com/giantswarm/observability-operator/pkg/config"
 )
 
 var managementClusterName = "dummy-cluster"
@@ -181,26 +179,27 @@ func TestGenerateAlloyConfig(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt // capture range variable
 		t.Run(tt.name, func(t *testing.T) {
 			ctx := context.Background()
 			// Create a dummy Service with minimal dependencies.
 			service := &Service{
 				OrganizationRepository: &MockOrganizationRepository{},
-				ManagementCluster: common.ManagementCluster{
-					InsecureCA: false,
-					Customer:   "dummy-customer",
-					Name:       "dummy-cluster",
-					Pipeline:   "dummy-pipeline",
-					Region:     "dummy-region",
-				},
-				MonitoringConfig: monitoring.Config{
-					WALTruncateFrequency: time.Minute,
-					QueueConfig: monitoring.QueueConfig{
-						Capacity:          ptr.To(30000),
-						MaxShards:         ptr.To(10),
-						MaxSamplesPerSend: ptr.To(150000),
-						SampleAgeLimit:    ptr.To("30m"),
+				Config: config.Config{
+					Cluster: config.ClusterConfig{
+						InsecureCA: false,
+						Customer:   "dummy-customer",
+						Name:       "dummy-cluster",
+						Pipeline:   "dummy-pipeline",
+						Region:     "dummy-region",
+					},
+					Monitoring: config.MonitoringConfig{
+						WALTruncateFrequency: time.Minute,
+						QueueConfig: config.QueueConfig{
+							Capacity:          &[]int{30000}[0],
+							MaxShards:         &[]int{10}[0],
+							MaxSamplesPerSend: &[]int{150000}[0],
+							SampleAgeLimit:    &[]string{"30m"}[0],
+						},
 					},
 				},
 			}